requirements.txt

mysql-connector-python
pycryptodome
